refactor(core): eliminate barrel files and consolidate utilities
## Changes

### 1. Removed All Barrel Files
- Deleted `packages/core/src/index.ts` (main barrel export)
- Renamed all remaining index.ts files to explicit module names:
  - `core-config/index.ts` → `core-config/coreConfig.ts`
  - `mutations/index.ts` → `mutations/MutationCoordinator.ts`
  - `summaries/index.ts` → `summaries/createSummary.ts`
  - `workflow/index.ts` → `workflow/verifyWorkflow.ts`

### 2. Enhanced Package Exports
Updated `packages/core/package.json` to include comprehensive wildcard exports:
- Added: `./evaluation/*`, `./improvement/*`, `./node/*`, `./tools/*`
- Updated: `./core-config` to point to explicit file path

### 3. Moved Stateless Utilities to @lucky/shared
Migrated pure, reusable utilities from core to shared package:
- `genShortId()` → `@lucky/shared` (crypto-based ID generation)
- `asArray()` → `@lucky/shared` (array coercion utility)
- `withDescriptions()` → `@lucky/shared` (Zod schema helper)

Updated both main and client exports in @lucky/shared
Fixed TypeScript compilation by adding DOM lib for crypto API

### 4. Updated All Imports
- Updated 53 files across the monorepo
- ~96 import statements changed to use explicit file paths
- Updated imports in core package, web app, and test files
- No more implicit index.ts resolution

## Benefits

- **Cleaner Architecture**: Explicit wildcard exports eliminate barrel file complexity
- **Better Modularity**: Pure utilities now centralized in shared package
- **Improved Reusability**: Stateless functions accessible across all packages
- **Reduced Coupling**: Core package focused on orchestration, not utilities
- **Type Safety**: Explicit imports improve IDE navigation and type checking

## Verification

✅ All builds pass
✅ TypeScript compilation succeeds
✅ Smoke tests pass
✅ No index.ts files remain in packages/core/src
