+++ apps/web/src/app/api/workflow/execute/route.ts
@@ -1,7 +1,7 @@
 import { requireAuth } from "@/lib/api-auth"
-import { genShortId } from "@lucky/core/utils/common/utils"
 import type { EvaluationText } from "@lucky/core/workflow/ingestion/ingestion.types"
 import type { WorkflowConfig } from "@lucky/core/workflow/schema/workflow.types"
+import { genShortId } from "@lucky/shared/client"
 import { type NextRequest, NextResponse } from "next/server"
 
 interface AsyncWorkflowExecution {
