+++ apps/web/src/app/api/workflow/verify/route.ts
@@ -1,6 +1,6 @@
 import { requireAuth } from "@/lib/api-auth"
 import { ensureCoreInit } from "@/lib/ensure-core-init"
-import { verifyWorkflowConfig } from "@lucky/core/utils/validation/workflow"
+import { verifyWorkflowConfig } from "@lucky/core/utils/validation/workflow/verifyWorkflow"
 import { clientWorkflowLoader } from "@lucky/core/workflow/setup/WorkflowLoader.client"
 import { type NextRequest, NextResponse } from "next/server"
