+++ apps/web/src/lib/core-init.ts
@@ -4,7 +4,7 @@
  * This file initializes the core module with the monorepo's runtime configuration.
  * Must be called before any core functionality is used.
  */
-import { initCoreConfig } from "@lucky/core/core-config"
+import { initCoreConfig } from "@lucky/core/core-config/coreConfig"
 import { CONFIG, MODELS, PATHS } from "@lucky/examples/settings/constants"
 
 let initialized = false
