+++ apps/web/src/trace-visualization/db/Workflow/retrieveWorkflow.ts
@@ -1,8 +1,7 @@
 "use server"
 import { supabase } from "@/lib/supabase"
-import { genShortId } from "@lucky/core/utils/common/utils"
 import { lgg } from "@lucky/core/utils/logging/Logger"
-import type { Tables, TablesInsert, TablesUpdate } from "@lucky/shared/client"
+import { type Tables, type TablesInsert, type TablesUpdate, genShortId } from "@lucky/shared/client"
 
 export const retrieveWorkflowInvocation = async (invocationId: string): Promise<Tables<"WorkflowInvocation">> => {
   const { data, error: WFInvocationError } = await supabase
