+++ packages/core/package.json
@@ -5,11 +5,15 @@
   "description": "Core workflow engine, evaluation, and evolution algorithms",
   "exports": {
     "./package.json": "./package.json",
-    "./core-config": "./src/core-config",
+    "./core-config": "./src/core-config/coreConfig.ts",
     "./workflow/*": "./src/workflow/*",
     "./utils/*": "./src/utils/*",
     "./messages/*": "./src/messages/*",
-    "./prompts/*": "./src/prompts/*"
+    "./prompts/*": "./src/prompts/*",
+    "./evaluation/*": "./src/evaluation/*",
+    "./improvement/*": "./src/improvement/*",
+    "./node/*": "./src/node/*",
+    "./tools/*": "./src/tools/*"
   },
   "scripts": {
     "dev": "vitest --watch --project unit",
