+++ packages/core/src/core-config/coreConfig.ts
