+++ packages/core/src/improvement/gp/Genome.ts
@@ -28,7 +28,6 @@ import { workflowConfigToGenome } from "@core/improvement/gp/resources/wrappers"
 import { SharedWorkflowPrompts } from "@core/prompts/workflowAnalysisPrompts"
 import type { FlowEvolutionMode } from "@core/types"
 import { truncater } from "@core/utils/common/llmify"
-import { genShortId } from "@core/utils/common/utils"
 import { lgg } from "@core/utils/logging/Logger"
 import { createWorkflowVersion, ensureWorkflowExists } from "@core/utils/persistence/workflow/registerWorkflow"
 import { getActiveModelNames } from "@core/utils/spending/functions"
@@ -36,7 +35,7 @@ import { Workflow } from "@core/workflow/Workflow"
 import type { EvaluationInput } from "@core/workflow/ingestion/ingestion.types"
 import { guard } from "@core/workflow/schema/errorMessages"
 import type { WorkflowConfig } from "@core/workflow/schema/workflow.types"
-import { R, type RS } from "@lucky/shared"
+import { R, type RS, genShortId } from "@lucky/shared"
 import { isNir } from "@lucky/shared"
 import type { IPersistence } from "@together/adapter-supabase"
 import type { GenomeEvaluationResults, WorkflowGenome } from "./resources/gp.types"
