+++ packages/core/src/improvement/gp/RunService.ts
@@ -8,9 +8,8 @@
 import type { EvolutionSettings, IterativeConfig } from "@core/improvement/gp/resources/evolution-types"
 import type { EvolutionContext } from "@core/improvement/gp/resources/types"
 import type { FlowEvolutionMode } from "@core/types"
-import { genShortId } from "@core/utils/common/utils"
 import { lgg } from "@core/utils/logging/Logger"
-import type { Tables, TablesInsert, TablesUpdate } from "@lucky/shared"
+import { type Tables, type TablesInsert, type TablesUpdate, genShortId } from "@lucky/shared"
 import { JSONN } from "@lucky/shared"
 import { isNir } from "@lucky/shared"
 import type { IEvolutionPersistence, IPersistence } from "@together/adapter-supabase"
