+++ packages/core/src/improvement/gp/operators/crossover/Crossover.ts
@@ -33,7 +33,7 @@ import { failureTracker } from "@core/improvement/gp/resources/tracker"
 import { workflowConfigToGenome } from "@core/improvement/gp/resources/wrappers"
 import { GENERALIZATION_LIMITS } from "@core/prompts/generalizationLimits"
 import { lgg } from "@core/utils/logging/Logger"
-import { verifyWorkflowConfig } from "@core/utils/validation/workflow"
+import { verifyWorkflowConfig } from "@core/utils/validation/workflow/verifyWorkflow"
 import { formalizeWorkflow } from "@core/workflow/actions/generate/formalizeWorkflow"
 import type { EvaluationInput } from "@core/workflow/ingestion/ingestion.types"
 import { R, type RS } from "@lucky/shared"
