+++ packages/core/src/improvement/gp/resources/wrappers.ts
@@ -5,7 +5,7 @@
 import type { WorkflowGenome } from "@core/improvement/gp/resources/gp.types"
 import type { EvolutionContext } from "@core/improvement/gp/resources/types"
 import { lgg } from "@core/utils/logging/Logger"
-import { verifyWorkflowConfig } from "@core/utils/validation/workflow"
+import { verifyWorkflowConfig } from "@core/utils/validation/workflow/verifyWorkflow"
 import type { EvaluationInput } from "@core/workflow/ingestion/ingestion.types"
 import type { WorkflowConfig } from "@core/workflow/schema/workflow.types"
 import { R, type RS } from "@lucky/shared"
