+++ packages/core/src/messages/api/vercel/vercelStepProcessor.ts
@@ -1,8 +1,7 @@
 import { calculateUsageCost } from "@core/messages/api/vercel/pricing/vercelUsage"
 import type { AgentSteps } from "@core/messages/pipeline/AgentStep.types"
-import { asArray } from "@core/utils/common/utils"
 import type { ModelName } from "@core/utils/spending/models.types"
-import { isNir } from "@lucky/shared"
+import { asArray, isNir } from "@lucky/shared"
 import { Tools } from "@lucky/shared"
 import type { StepResult, ToolCallPart, ToolSet } from "ai"
