+++ packages/core/src/messages/pipeline/agentStepLoop/MultiStepLoopV3.ts
@@ -20,7 +20,7 @@ import { sendAI } from "@core/messages/api/sendAI/sendAI"
 import type { ProcessedResponse } from "@core/messages/api/vercel/processResponse.types"
 import { responseToAgentSteps } from "@core/messages/api/vercel/responseToAgentSteps"
 import { selectToolStrategyV3 } from "@core/messages/pipeline/selectTool/selectToolStrategyV3"
-import { generateSummaryFromUnknownData } from "@core/messages/summaries"
+import { generateSummaryFromUnknownData } from "@core/messages/summaries/createSummary"
 import { makeLearning } from "@core/prompts/makeLearning"
 import { llmify } from "@core/utils/common/llmify"
 import { lgg } from "@core/utils/logging/Logger"
