+++ packages/core/src/node/schemas/agentWithTools.ts
@@ -1,6 +1,6 @@
 import { CONFIG } from "@core/core-config/compat"
 import { AGENT_KEY_EXPLANATIONS, baseWorkflowNodeConfigShape } from "@core/node/schemas/improvementSchema"
-import { withDescriptions } from "@core/utils/zod/withDescriptions"
+import { withDescriptions } from "@lucky/shared"
 import { ACTIVE_CODE_TOOL_NAMES, ACTIVE_MCP_TOOL_NAMES } from "@lucky/tools/client"
 import type { z } from "zod"
