+++ packages/core/src/utils/fs/fileSaver.ts
@@ -6,7 +6,7 @@
 import fs from "node:fs"
 import path from "node:path"
 import { fileURLToPath } from "node:url"
-import { getCorePaths } from "@core/core-config"
+import { getCorePaths } from "@core/core-config/coreConfig"
 import { JSONN } from "@lucky/shared"
 import type { CodeToolResult } from "@lucky/tools"
 import { nanoid } from "nanoid"
