+++ packages/core/src/utils/persistence/memory/SupabaseStore.ts
@@ -1,4 +1,4 @@
-import { generateSummaryFromUnknownData } from "@core/messages/summaries"
+import { generateSummaryFromUnknownData } from "@core/messages/summaries/createSummary"
 import type { ContextFileInfo } from "@core/tools/context/contextStore.types"
 import { supabase } from "@core/utils/clients/supabase/client"
 import { lgg } from "@core/utils/logging/Logger"
