+++ packages/core/src/workflow/Workflow.ts
@@ -5,12 +5,11 @@ import { type PrepareProblemMethod, prepareProblem } from "@core/improvement/beh
 import type { EvolutionContext } from "@core/improvement/gp/resources/types"
 import { WorkFlowNode } from "@core/node/WorkFlowNode"
 import type { WorkflowFile } from "@core/tools/context/contextStore.types"
-import { genShortId } from "@core/utils/common/utils"
 import { lgg } from "@core/utils/logging/Logger"
 import { persistWorkflow } from "@core/utils/persistence/file/resultPersistence"
 import { type ContextStore, createContextStore } from "@core/utils/persistence/memory/ContextStore"
 import type { ModelName } from "@core/utils/spending/models.types"
-import { verifyWorkflowConfig, verifyWorkflowConfigStrict } from "@core/utils/validation/workflow"
+import { verifyWorkflowConfig, verifyWorkflowConfigStrict } from "@core/utils/validation/workflow/verifyWorkflow"
 import { zodToJson } from "@core/utils/zod/zodToJson"
 import { formalizeWorkflow } from "@core/workflow/actions/generate/formalizeWorkflow"
 import { type SimplifyOptions, workflowToString } from "@core/workflow/actions/generate/workflowToString"
@@ -20,6 +19,7 @@ import type { AggregateEvaluationResult, RunResult } from "@core/workflow/runner
 import { ensure, guard, throwIf } from "@core/workflow/schema/errorMessages"
 import { hashWorkflow } from "@core/workflow/schema/hash"
 import type { WorkflowConfig, WorkflowNodeConfig } from "@core/workflow/schema/workflow.types"
+import { genShortId } from "@lucky/shared"
 import type { RS } from "@lucky/shared"
 import { R } from "@lucky/shared"
 import { INACTIVE_TOOLS } from "@lucky/tools"
