+++ packages/core/src/workflow/runner/invokeWorkflow.ts
@@ -13,16 +13,15 @@
  */
 
 import { CONFIG } from "@core/core-config/compat"
-import { genShortId } from "@core/utils/common/utils"
 import { lgg } from "@core/utils/logging/Logger"
 import { obs } from "@core/utils/observability/obs"
 import { SpendingTracker } from "@core/utils/spending/SpendingTracker"
-import { verifyWorkflowConfigStrict } from "@core/utils/validation/workflow"
+import { verifyWorkflowConfigStrict } from "@core/utils/validation/workflow/verifyWorkflow"
 import { Workflow } from "@core/workflow/Workflow"
 import { needsEvaluation } from "@core/workflow/ingestion/ingestion.types"
 import type { WorkflowConfig } from "@core/workflow/schema/workflow.types"
 import { loadFromDSL, loadFromDatabase, loadFromFile } from "@core/workflow/setup/WorkflowLoader"
-import { JSONN } from "@lucky/shared"
+import { JSONN, genShortId } from "@lucky/shared"
 import { R, type RS } from "@lucky/shared"
 import { isNir } from "@lucky/shared"
 import type { InvocationInput, InvokeWorkflowResult, RunResult } from "./types"
