+++ packages/core/src/workflow/runner/queueRun.ts
@@ -4,7 +4,7 @@ import { CONFIG } from "@core/core-config/compat"
 import type { AggregatedPayload, MessageType } from "@core/messages/MessagePayload"
 import { WorkflowMessage } from "@core/messages/WorkflowMessage"
 import type { AgentSteps } from "@core/messages/pipeline/AgentStep.types"
-import type { InvocationSummary } from "@core/messages/summaries"
+import type { InvocationSummary } from "@core/messages/summaries/createSummary"
 import { lgg } from "@core/utils/logging/Logger"
 import { updateWorkflowMemory } from "@core/utils/persistence/workflow/updateNodeMemory"
 import { getNodeRole } from "@core/utils/validation/workflow/verifyHierarchical"
