+++ packages/core/src/workflow/schema/workflowSchema.ts
@@ -6,8 +6,8 @@ import { mapModelNameToEasyName } from "@core/prompts/explainAgents"
 import { MemorySchemaOptional } from "@core/utils/memory/memorySchema"
 import type { AnyModelName, ModelName } from "@core/utils/spending/models.types"
 import { ACTIVE_MODEL_NAMES } from "@core/utils/spending/pricing"
-import { withDescriptions } from "@core/utils/zod/withDescriptions"
 import type { WorkflowConfig, WorkflowNodeConfig } from "@core/workflow/schema/workflow.types"
+import { withDescriptions } from "@lucky/shared"
 import { ACTIVE_CODE_TOOL_NAMES_WITH_DEFAULT, ACTIVE_MCP_TOOL_NAMES } from "@lucky/tools/client"
 
 export const WorkflowNodeConfigSchema = z.object({
