+++ packages/core/src/workflow/setup/WorkflowLoader.ts
@@ -1,8 +1,8 @@
 import { CONFIG, PATHS, isLoggingEnabled } from "@core/core-config/compat"
 import { mkdirIfMissing } from "@core/utils/common/files"
 import { lgg } from "@core/utils/logging/Logger"
-import { verifyWorkflowConfig } from "@core/utils/validation/workflow"
 import { isValidToolInformation } from "@core/utils/validation/workflow/toolInformation"
+import { verifyWorkflowConfig } from "@core/utils/validation/workflow/verifyWorkflow"
 import type { WorkflowConfig, WorkflowNodeConfig } from "@core/workflow/schema/workflow.types"
 import { WorkflowConfigSchema, WorkflowConfigSchemaDisplay } from "@core/workflow/schema/workflowSchema"
 import type { CodeToolName } from "@lucky/tools"
