+++ packages/shared/src/utils/common/id.ts
@@ -1,0 +1,8 @@
+/**
+ * Generate a short ID using crypto.randomUUID()
+ * Returns first 8 characters of UUID without hyphens
+ */
+export const genShortId = () => {
+  const fullUuid = globalThis.crypto.randomUUID().replace(/-/g, "")
+  return fullUuid.substring(0, 8)
+}
