+++ packages/shared/tsconfig.build.json
@@ -3,7 +3,7 @@
     "module": "ESNext",
     "moduleResolution": "bundler",
     "target": "ES2020",
-    "lib": ["ESNext"],
+    "lib": ["ESNext", "DOM"],
     "types": ["node"],
     "rootDir": "./src",
     "outDir": "./dist",
